Phase 4.1a: extract MarkdownRenderer from HelpPanel + use in BodyTab
The hand-rolled markdown parser previously lived inline in HelpPanel.tsx.
Phase 4.1 needs the same renderer for the agents/skills Body tab, so
the parse + inline-format logic moves to src/components/ui/MarkdownRenderer.tsx
and the help-system slug routing becomes an opt-in onLinkClick prop.

HelpPanel keeps its slug-routing behavior by passing a delegate that
recognizes *.md hrefs; BodyTab omits the prop so links are inert.

15 new tests cover block-level (headings, paragraphs, fenced code, lists,
tables) and inline (bold, code, link, mixed) parsing.
